now handles very large files using 64bit ints

+++ kmer_db.h
@@ -2,6 +2,7 @@
 #define KMER_DB_H
 
 #include <sys/stat.h>
+#include <stdint.h>
 
 #include "kmer_record.h"
 //typedef struct KmerRecord_struct *KmerRecordPtr;
@@ -10,14 +11,14 @@ typedef struct KmerDb_struct{
   int mapped_file_handle;
   char *mmapfile;
   KmerRecord  *kmer_record;
-  int num_records;
+  uint64_t num_records;
   int record_size;
   struct stat statbuf;
 }KmerDb;
 
 KmerDb *newKmerDb(char *file);
 void freeKmerDb(KmerDb *kdb);
-KmerRecord *getRecordFromDb(KmerDb *db, int idx);
+KmerRecord *getRecordFromDb(KmerDb *db, uint64_t idx);
 KmerRecord *findRecordFromDb(KmerDb *db, const char *kmer);
 
 #endif
